refactor (numberGenerator) : 같은 역할 리포지터리로 묶기
