+++ src/main/kotlin/Application.kt
@@ -2,7 +2,7 @@ package racingcar
 
 import racingcar.controller.GameController
 import racingcar.domain.RaceManager
-import racingcar.domain.RandomNumberGenerator
+import racingcar.racingcar.domain.numbergenerator.RandomNumberGenerator
 
 fun main() {
     GameController(RaceManager(RandomNumberGenerator())).runGame()
