+++ src/main/kotlin/racingcar/domain/numbergenerator/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.racingcar.domain.numbergenerator
 
 interface NumberGenerator {
     fun generateNumber(minNumber: Int, maxNumber: Int): Int
