+++ src/main/kotlin/racingcar/domain/numbergenerator/RandomNumberGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.racingcar.domain.numbergenerator
 
 class RandomNumberGenerator : NumberGenerator {
     override fun generateNumber(minNumber: Int, maxNumber: Int): Int {
