+++ src/test/kotlin/racingcar/domain/CarFactoryTest.kt
@@ -3,6 +3,7 @@ package racingcar.domain
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import racingcar.racingcar.domain.CarFactory
+import racingcar.racingcar.domain.numbergenerator.RandomNumberGenerator
 
 class CarFactoryTest {
     @Test
